correct sizes for magstripe read, but...
Has an issue where the read needs to occur twice for some reason? As in, first read it won't come back with any data, but if the program is run again (without swiping a second time), then the data is read. Need to review Wireshark capture and figure out what's going wrong.